requirements.txt
requests
Pillow
numpy
opencv-python

Flask
flask-restful

pymongo
